% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_p.R
\name{report_p}
\alias{report_p}
\title{report_p}
\usage{
report_p(p, digits = 3)
}
\arguments{
\item{p}{a number}

\item{digits}{the number of digits}
}
\value{
Returns the value of \code{arg1}
}
\description{
report_p returns p in APA style with the number of digits provides.
}
\examples{

report_p(0.346, 3) # returns p = .346

}
