% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_p.R
\name{report_p}
\alias{report_p}
\title{report_p}
\usage{
report_p(p, digits = 3)
}
\arguments{
\item{p}{A value to return}
}
\value{
Returns the p-value \code{p} in APA style
}
\description{
\code{report_p} reports p-values in APA style
}
\examples{

report_p(1) # returns 1000

}
