% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyFunction.R, R/report_p.R
\name{report_p}
\alias{report_p}
\title{report_p}
\usage{
report_p(p, digits = 3)

report_p(p, digits = 3)
}
\arguments{
\item{arg1}{A value to return}
}
\value{
Returns the value of \code{arg1}

Returns the value of \code{arg1}
}
\description{
\code{report_p} does something.

\code{report_p} does something.
}
\examples{

myfunction(1) # returns 1


report_p(0.2) # returns "p=.200"


}
