% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_p.R
\name{report_p}
\alias{report_p}
\title{report_p}
\usage{
report_p(p, digits = 3)
}
\arguments{
\item{p}{A value to evaluate}

\item{digits}{is significant figures}
}
\value{
Returns the value of \code{p_string}
}
\description{
\code{report_p} does something good.
}
\examples{

report_p(1) # returns 1

}
